requirements.txt
httpx>=0.27.0
web3>=6.15.0
eth-account>=0.11.0
Pillow>=10.3.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
